Simplify code in launch and debug
