+++ src/client/common/vscodeApis/workspaceApis.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { ConfigurationScope, workspace, WorkspaceConfiguration, WorkspaceFolder } from 'vscode';
+import { ConfigurationScope, workspace, WorkspaceConfiguration, WorkspaceEdit, WorkspaceFolder } from 'vscode';
 import { Resource } from '../types';
 
 export function getWorkspaceFolders(): readonly WorkspaceFolder[] | undefined {
@@ -18,4 +18,8 @@ export function getWorkspaceFolderPaths(): string[] {
 
 export function getConfiguration(section?: string, scope?: ConfigurationScope | null): WorkspaceConfiguration {
     return workspace.getConfiguration(section, scope);
+}
+
+export function applyEdit(edit: WorkspaceEdit): Thenable<boolean> {
+    return workspace.applyEdit(edit);
 }
