+++ src/client/debugger/extension/configuration/resolvers/base.ts
@@ -8,6 +8,10 @@ import * as path from 'path';
 import { CancellationToken, DebugConfiguration, Uri, WorkspaceFolder } from 'vscode';
 import { IConfigurationService } from '../../../../common/types';
 import { getOSType, OSType } from '../../../../common/utils/platform';
+import {
+    getWorkspaceFolder as getVSCodeWorkspaceFolder,
+    getWorkspaceFolders,
+} from '../../../../common/vscodeApis/workspaceApis';
 import { IInterpreterService } from '../../../../interpreter/contracts';
 import { sendTelemetryEvent } from '../../../../telemetry';
 import { EventName } from '../../../../telemetry/constants';
@@ -16,7 +20,6 @@ import { AttachRequestArguments, DebugOptions, LaunchRequestArguments, PathMappi
 import { PythonPathSource } from '../../types';
 import { IDebugConfigurationResolver } from '../types';
 import { resolveVariables } from '../utils/common';
-import { getWorkspaceFolder as getVSCodeWorkspaceFolder, getWorkspaceFolders } from '../utils/workspaceFolder';
 import { getProgram } from './helper';
 
 @injectable()
