--- src/client/debugger/extension/configuration/utils/workspaceFolder.ts
@@ -1,17 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-'use strict';
-
-import * as vscode from 'vscode';
-
-export function getWorkspaceFolder(uri: vscode.Uri): vscode.WorkspaceFolder | undefined {
-    return vscode.workspace.getWorkspaceFolder(uri);
-}
-
-export function getWorkspaceFolders(): readonly vscode.WorkspaceFolder[] | undefined {
-    return vscode.workspace.workspaceFolders;
-}
